Moved setting up buttons to base activity
